backend/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
dj-database-url==1.2.0
Django==3.2.16
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simpleJWT==5.2.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.3.0
